Created reusable dt instance
